Agregada una variable para contar la cantidad de evaluaciones de las funciones objetivo. Además, se arreglaron errores en la función DB(K) y CS(K) en el archivo Metaheuristic.cpp

+++ src/helpers.h
@@ -51,7 +51,7 @@ using namespace std;
 /**
  * Tiempo de ejecución.
  */
-extern float runtime;
+extern double runtime;
 
 ///////////////////////////////////////////
 // Rquisitos genéricos
